PUSH DE HUGO - ahora el heuristico no da caca

+++ LocalSearch/src/IA/Bicing/BicingHeuristic.java
@@ -4,7 +4,7 @@ import aima.search.framework.HeuristicFunction;
 
 public class BicingHeuristic implements HeuristicFunction {
 	public double getHeuristicValue(Object state){
-        int h = -((BicingBoard) state).get_heur1();
+        double h = - ((BicingBoard) state).get_heur1();
         return h;
     }
 }
